requirements.txt
google-generativeai
python-dotenv
requests
pathlib2